+++ .idea/workspace.xml
@@ -5,16 +5,10 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="b19a05db-8e42-430a-9bdc-42edde6671c4" name="변경" comment="">
-      <change beforePath="$PROJECT_DIR$/.idea/.gitignore" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/OkiOki.iml" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/modules.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/modules.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/uiDesigner.xml" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/OkiOki.iml" beforeDir="false" afterPath="$PROJECT_DIR$/OkiOki.iml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/Main.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/Main.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/cart/Bucket.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/cart/Bucket.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/cart/Main.java" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/cart/Options.java" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/cart/temp_bucket01.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/cart/temp_bucket01.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/cart/temp_bucket01.java" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -64,7 +58,7 @@
       <option name="number" value="Default" />
       <option name="presentableId" value="Default" />
       <updated>1730357291166</updated>
-      <workItem from="1730357292490" duration="2244000" />
+      <workItem from="1730357292490" duration="3419000" />
     </task>
     <servers />
   </component>
