+++ src/cart/Bucket.java
@@ -1,95 +1,117 @@
 package cart;
 
-import order.OrderDetails;
-
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Scanner;
 
 public class Bucket {
-    private ArrayList<OrderDetails> cartItems = new ArrayList<>();
-    private Scanner reader;
+    private ArrayList<Map<String, Object>> cart;
+    private Scanner scanner;
 
-    public Bucket(Scanner reader) {
-        this.reader = reader;
+    public Bucket(){
+        this.cart = new ArrayList<>();
+        this.scanner = new Scanner(System.in);
     }
 
-    // 주문 추가 메서드
-    public void addToBucket(OrderDetails order) {
-        cartItems.add(order);
-        System.out.println(order.getMenu() + "가 장바구니에 추가되었습니다.");
+    public void add_order(Map<String, Object> orderInfo){
+        Map<String, Object> newOrder = new HashMap<>(orderInfo);
+        cart.add(newOrder);
     }
 
-    // 현재 주문 상태 표시
-    public void displayCurrentItem(OrderDetails order) {
-        System.out.println("메뉴: " + order.getMenu() + " (온도: " +
-                (order.getTemperature() == 0 ? "HOT" : "ICE") + ")");
-        System.out.println("옵션:");
-        for (String option : order.getOptions().keySet()) {
-            System.out.println("  - " + option);
+    public void display_bucket(){
+        for (int i = 0; i < cart.size(); i++) {
+            Map<String, Object> order = cart.get(i);
+            System.out.println("-----");
+            System.out.println((i + 1) + ". 메뉴: " + order.get("menu"));
+            System.out.println("HOT/ICE: " + order.get("temperature"));
+            System.out.println("선택된 옵션: " + order.get("options"));
+            System.out.println("총 가격: " + order.get("price"));
+            System.out.println("-----");
         }
-        System.out.println("총 가격: " + order.calculateTotalPrice() + "원");
     }
 
-    // 장바구니 표시
-    public void displayBucket() {
-        if (cartItems.isEmpty()) {
-            System.out.println("장바구니가 비었습니다.💨");
+    public void delete_order() {
+        display_bucket();
+        if (cart.isEmpty()) {
+            System.out.println("장바구니가 비어 있습니다.");
             return;
         }
 
-        System.out.println("=== 🧺장바구니🧺 ===");
-        for (int i = 0; i < cartItems.size(); i++) {
-            OrderDetails order = cartItems.get(i);
-            System.out.println((i + 1) + ". " + order.getMenu() + " (" +
-                    (order.getTemperature() == 0 ? "HOT" : "ICE") + ") - " +
-                    order.calculateTotalPrice() + "원");
-        }
-    }
-
-    // 장바구니 항목 수정
-    public void modifyBucket() {
-        displayBucket();
-        if (cartItems.isEmpty()) return;
-
-        System.out.print("수정할 음료의 번호를 입력하세요: ");
-        String input = reader.nextLine().trim();
-        int itemNum;
-
-        try {
-            itemNum = Integer.parseInt(input);
-            if (itemNum > 0 && itemNum <= cartItems.size()) {
-                OrderDetails order = cartItems.get(itemNum - 1);
-                System.out.println(order.getMenu() + "가 선택되었습니다. 새 이름을 입력하세요:");
-                String newMenuName = reader.nextLine().trim();
-                // order 객체에 새로운 메뉴 이름 설정 (OrderDetails에 setMenu 메서드 필요)
-                System.out.println("장바구니 수정 성공");
-            } else {
-                System.out.println("잘못된 번호입니다.");
-            }
-        } catch (NumberFormatException e) {
-            System.out.println("잘못된 입력입니다.");
-        }
-    }
-
-    // 장바구니 항목 삭제
-    public void removeFromBucket() {
-        displayBucket();
-        if (cartItems.isEmpty()) return;
-
         System.out.print("삭제할 음료 번호를 입력하세요: ");
-        String input = reader.nextLine().trim();
-        int itemNum;
+        int orderIndex = scanner.nextInt() - 1;
+        scanner.nextLine(); // 개행 제거
 
-        try {
-            itemNum = Integer.parseInt(input);
-            if (itemNum > 0 && itemNum <= cartItems.size()) {
-                cartItems.remove(itemNum - 1);
-                System.out.println("음료가 삭제되었습니다.");
-            } else {
-                System.out.println("잘못된 번호입니다.");
-            }
-        } catch (NumberFormatException e) {
-            System.out.println("잘못된 입력입니다.");
+        if (orderIndex >= 0 && orderIndex < cart.size()) {
+            cart.remove(orderIndex);
+            System.out.println("음료가 삭제되었습니다.");
+        } else {
+            System.out.println("잘못된 번호입니다.");
         }
+    }
+
+    public void modify_options() {
+        display_bucket();
+        if (cart.isEmpty()) {
+            System.out.println("장바구니가 비어 있습니다.");
+            return;
+        }
+
+        System.out.print("옵션을 수정할 음료 번호를 입력하세요: ");
+        int orderIndex = scanner.nextInt() - 1;
+        scanner.nextLine(); // 개행 제거
+
+        if (orderIndex >= 0 && orderIndex < cart.size()) {
+            Map<String, Object> order = cart.get(orderIndex);
+            Map<String, String> options = (Map<String, String>) order.get("options");
+
+            System.out.print("수정할 옵션 이름을 입력하세요: ");
+            String optionName = scanner.nextLine();
+
+            if (options.containsKey(optionName)) {
+                System.out.print("새 옵션 값을 입력하세요: ");
+                String newOptionValue = scanner.nextLine();
+                options.put(optionName, newOptionValue);
+
+                System.out.println("옵션이 수정되었습니다.");
+            } else {
+                System.out.println("해당 옵션이 없습니다.");
+            }
+        } else {
+            System.out.println("잘못된 번호입니다.");
+        }
+    }
+
+    public void delete_options() {
+        display_bucket();
+        if (cart.isEmpty()) {
+            System.out.println("장바구니가 비어 있습니다.");
+            return;
+        }
+
+        System.out.print("옵션을 삭제할 음료 번호를 입력하세요: ");
+        int orderIndex = scanner.nextInt() - 1;
+        scanner.nextLine(); // 개행 제거
+
+        if (orderIndex >= 0 && orderIndex < cart.size()) {
+            Map<String, Object> order = cart.get(orderIndex);
+            Map<String, String> options = (Map<String, String>) order.get("options");
+
+            System.out.print("삭제할 옵션 이름을 입력하세요: ");
+            String optionName = scanner.nextLine();
+
+            if (options.containsKey(optionName)) {
+                options.remove(optionName);
+                System.out.println("옵션이 삭제되었습니다.");
+            } else {
+                System.out.println("해당 옵션이 없습니다.");
+            }
+        } else {
+            System.out.println("잘못된 번호입니다.");
+        }
+    }
+
+    public ArrayList<Map<String, Object>> getCart() {
+        return cart;
     }
 }
